Crate emit_elasticsearch

emit_elasticsearch

Log events with the emit structured logger to Elasticsearch.

This is a lightweight Collector that indexes events in a timestamped index with a _bulk command. The template format can be changed by specifying an IndexTemplate, but the log type will always be emitlog. Logged events should play nice with kibana out-of-the-box.

Usage

Reference emit_elasticsearch in your Cargo.toml:

[dependencies]
emit = "*"
emit_elasticsearch = "*"

Log to Elasticsearch by adding an ElasticCollector to your pipeline:

#[macro_use]
extern crate emit;
extern crate emit_elasticsearch;
 
use emit::PipelineBuilder;
use emit_elasticsearch::prelude::*;
 
let pipeline = PipelineBuilder::new()
    .write_to(
        ElasticCollector::default()
        .send_template().unwrap())
    .init();

The call to .send_template is optional, and will ensure you have an index_template in Elasticsearch to map events correctly.

Events are formatted using the RenderedJsonFormatter. So a warning log with a template "The number is {number}" will produce a result like the following:

{
    "@t": "2014-07-08T09:10:11.000Z",
    "@m": "The number is 42",
    "@i": "ae9bf784",
    "@l": "WARN",
    "number":42
}

Modules

prelude

Structs

ElasticCollector
Log collector for Elasticsearch.

IndexTemplate
Template for naming log indices.

Constants

DEFAULT_TEMPLATE_FORMAT
The default date format appended to the prefix for log indices.

DEFAULT_TEMPLATE_PREFIX
The default index template prefix for log indices.

LOCAL_SERVER_URL
The default address for a local Elasticsearch node.

TYPENAME
The value of _type used for indexed logs.