Struct emit_elasticsearch::IndexTemplate

pub struct IndexTemplate {
    // some fields omitted
}

Template for naming log indices.

The index name consists of a prefix and a date format (chrono compatible). The default IndexTemplate produces index names like 'emitlog-20160501'.

Examples

An index template that produces index names like 'logs-2016-05-01':

let template = IndexTemplate::new("logs-", "%Y-%m-%d");

Use the date_format parameter to control how the event date is resolved to an index name. For example, an index template that produces a unique index name for each month:

let template = IndexTemplate::new("emitlog-", "%Y%m");

The ElasticCollector will append * to the end of your IndexTemplate.prefix when inserting an in Elasticsearch. So the above examples will both use "emitlog-*" as the template value in the call to index_template.

Methods

impl IndexTemplate

fn new<I>(prefix: I, date_format: I) -> IndexTemplate where I: Into<String>

Create an index template with the given prefix and date_format.

fn index(&self, date: &DateTime<UTC>) -> String

Get the index name for a given timestamp.

This will use the date_format to format the timestamp and append it to the prefix. So a template with a prefix of "testlog-" and format of "%Y%m%d" will produce index names like the following:

"testlog-20140708"